1、更新数据库文件 2、编写作者表swagger文件 3、生成持久层代码 4、生成接口代码 5、编写业务代码

+++ src/main/java/org/garen/mc/mybatis/mapper/CommonMapper.java
@@ -17,7 +17,7 @@ public interface CommonMapper<T,Q,PK extends Serializable>  {
     /**
      * 根据条件查询记录总数
      */
-    int countByExample(Q example);
+    long countByExample(Q example);
 
     /**
      * 根据条件删除记录
